Support MySQL by treating usernames as case-insensitive
MySQL performs case-insensitive search by default, whereas PostgreSQL requires the
`ILIKE` keyword to force a case-insensitive search. The solution applied here is to
treat usernames as case-insensitive by automatically applying `.toLowerCase()` to
the username provided when creating the user. This means that `andrew` and `Andrew`
will now be considered the same username, which is probably preferable anyway as
it avoids confusion (e.g.: 'did I type it with a capital 'A' when I signed up?').